--- src/main/java/kr/where/backend/exception/customException/AdminLoginFailException.java
@@ -1,26 +1,0 @@
-package kr.where.backend.exception.customException;
-
-import kr.where.backend.utils.response.ResponseMsg;
-import kr.where.backend.utils.response.StatusCode;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * <pre>
- *     관리자 로그인 시 id - pwd 불일치 시 발생
- *     error code: 401
- * </pre>
- * @see AdminService#adminLogin(String, String)
- * @since 1.0
- * @author hyunjcho
- */
-@Getter
-@Slf4j
-public class AdminLoginFailException extends RuntimeException{
-	private int errorCode;
-
-	public AdminLoginFailException() {
-		super(ResponseMsg.ADMIN_FAIL);
-		this.errorCode = StatusCode.UNAUTHORIZED;
-	}
-}
