+++ src/main/java/kr/where/backend/group/exception/GroupMemberException.java
@@ -1,6 +1,6 @@
 package kr.where.backend.group.exception;
 
-import kr.where.backend.suhwparkException.CustomException;
+import kr.where.backend.exception.CustomException;
 
 public class GroupMemberException extends CustomException {
     public GroupMemberException(final GroupMemberErrorCode groupMemberErrorCode) {
